Moves UpsertUsersObserver specs into a single file
